[TEST] Added cargo image e2e tests
