+++ tests/version.rs
@@ -1,13 +1,8 @@
 pub mod common;
 
-#[cfg(test)]
-pub mod version {
-  use super::common;
-
-  #[ntex::test]
-  async fn exec_version() -> Result<(), common::TestError> {
-    let output = common::spawn_cli(vec!["version"]).await?;
-    assert!(output.status.success());
-    Ok(())
-  }
+#[ntex::test]
+async fn exec() -> Result<(), common::TestError> {
+  let output = common::spawn_cli(vec!["version"]).await?;
+  assert!(output.status.success());
+  Ok(())
 }
